Finish face annotation
